Insert, delete and participate added

+++ app/src/main/java/com/example/joinme/data/entities/Activity.kt
@@ -15,9 +15,10 @@ import androidx.room.PrimaryKey
 data class Activity (
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val title: String = "",
-    val creatorId: Int,
+    val creatorId: Long,
     val date: String = "",
     val time: String = "",
+    //var currentParticipants: Int = 0,
     val maxParticipants: Int = 0,
     val location: String = "",
     )
