+++ app/src/main/java/com/example/joinme/data/entities/User.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity (tableName = "users")
 data class User (
-    @PrimaryKey(autoGenerate = true) val userId: Int = 0,
+    @PrimaryKey(autoGenerate = true) val userId: Long = 0,
     val username: String = "",
     val email: String = "",
 )
